backend-fastapi/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.76.0
supabase==2.15.0
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.11.3
python-jose==3.3.0
passlib==1.7.4
